重构代码，删除RHI

+++ engine/src/runtime/include/common/math/literal.hpp
@@ -10,7 +10,7 @@ namespace taixu::literal {
 
 static constexpr uint8_t UC_MAX{255};
 
-constexpr uint8_t operator"" _uc(const unsigned long long value) {
+constexpr uint8_t operator"" _uc(const uint64_t value) {
     if (value > UC_MAX) { return UC_MAX; }
     return static_cast<uint8_t>(value);
 }
